+++ infoscience_exports/exports/versions.py
@@ -3,14 +3,14 @@
 
 # the release comes from git and should not be modified
 # => read-only
-_release = '0.4.4-14-g63ec68f'
+_release = '0.4.6-44-g1e9a3ba'
 
 # you can set the next version number manually
 # if you do not, the system will make sure that version > release
 # => read-write, >_release
-_version = '0.4.5'
+_version = '0.5.0'
 
 # the build number will generate conflicts on each PR merge
 # just keep yours every time
 # => read-only
-_build = '63ec68f83065413739aced244c7852b575ea4785'
+_build = '1e9a3ba06decda25003079d0b1a3770f54eb13da'