+++ CHANGELOG.md
@@ -1,10 +1,34 @@
 # Change Log
 
-## [qa-release](https://github.com/epfl-idevelop/infoscience-exports/tree/qa-release) (2018-04-13)
-[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.5...qa-release)
+## [0.5.0](https://github.com/epfl-idevelop/infoscience-exports/tree/0.5.0) (2018-04-26)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/qa-release...0.5.0)
+
+## [qa-release](https://github.com/epfl-idevelop/infoscience-exports/tree/qa-release) (2018-04-26)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.7...qa-release)
+
+**Merged pull requests:**
+
+- Migration subset [\#105](https://github.com/epfl-idevelop/infoscience-exports/pull/105) ([jdelasoie](https://github.com/jdelasoie))
+- Change OnetoOne Legacy Export to OnetoMany [\#104](https://github.com/epfl-idevelop/infoscience-exports/pull/104) ([jdelasoie](https://github.com/jdelasoie))
+
+## [0.4.7](https://github.com/epfl-idevelop/infoscience-exports/tree/0.4.7) (2018-04-25)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.6...0.4.7)
+
+**Merged pull requests:**
+
+- Selective migration [\#103](https://github.com/epfl-idevelop/infoscience-exports/pull/103) ([jdelasoie](https://github.com/jdelasoie))
+
+## [0.4.6](https://github.com/epfl-idevelop/infoscience-exports/tree/0.4.6) (2018-04-24)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.5...0.4.6)
+
+**Merged pull requests:**
+
+- Url migration process [\#102](https://github.com/epfl-idevelop/infoscience-exports/pull/102) ([jdelasoie](https://github.com/jdelasoie))
+- Removed duplicate call to domain when clipboarding [\#100](https://github.com/epfl-idevelop/infoscience-exports/pull/100) ([dragonleman](https://github.com/dragonleman))
+- RANGE\_DISPLAY in env variables [\#99](https://github.com/epfl-idevelop/infoscience-exports/pull/99) ([dragonleman](https://github.com/dragonleman))
 
 ## [0.4.5](https://github.com/epfl-idevelop/infoscience-exports/tree/0.4.5) (2018-04-13)
-[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/prod-release...0.4.5)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.4...0.4.5)
 
 **Merged pull requests:**
 
@@ -15,9 +39,6 @@
 - Titres avec Liste\(s\) de publications [\#94](https://github.com/epfl-idevelop/infoscience-exports/pull/94) ([dragonleman](https://github.com/dragonleman))
 - Added collectstatic an makemessages when deploying [\#93](https://github.com/epfl-idevelop/infoscience-exports/pull/93) ([dragonleman](https://github.com/dragonleman))
 - User creation for admins [\#92](https://github.com/epfl-idevelop/infoscience-exports/pull/92) ([jdelasoie](https://github.com/jdelasoie))
-
-## [prod-release](https://github.com/epfl-idevelop/infoscience-exports/tree/prod-release) (2018-04-06)
-[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.4...prod-release)
 
 ## [0.4.4](https://github.com/epfl-idevelop/infoscience-exports/tree/0.4.4) (2018-04-06)
 [Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.4.3...0.4.4)
@@ -98,7 +119,7 @@
 - Remove any usages or dependencies on Django Rest Framework [\#49](https://github.com/epfl-idevelop/infoscience-exports/pull/49) ([jdelasoie](https://github.com/jdelasoie))
 
 ## [0.3.5](https://github.com/epfl-idevelop/infoscience-exports/tree/0.3.5) (2018-03-13)
-[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.3.4...0.3.5)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/prod-release...0.3.5)
 
 **Merged pull requests:**
 
@@ -106,6 +127,9 @@
 - FullText: only one link to redirect to a pdf or a directory with several media [\#47](https://github.com/epfl-idevelop/infoscience-exports/pull/47) ([dragonleman](https://github.com/dragonleman))
 - Corrected group\_by year [\#46](https://github.com/epfl-idevelop/infoscience-exports/pull/46) ([dragonleman](https://github.com/dragonleman))
 - Translations: locale folders [\#45](https://github.com/epfl-idevelop/infoscience-exports/pull/45) ([dragonleman](https://github.com/dragonleman))
+
+## [prod-release](https://github.com/epfl-idevelop/infoscience-exports/tree/prod-release) (2018-03-09)
+[Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.3.4...prod-release)
 
 ## [0.3.4](https://github.com/epfl-idevelop/infoscience-exports/tree/0.3.4) (2018-03-09)
 [Full Changelog](https://github.com/epfl-idevelop/infoscience-exports/compare/0.3.3...0.3.4)